Mudando uma série de coisas na modelagem do banco de dados e na adição de uma nova coluna de usuários para salvar melhor as senhas dos usuários

+++ database/migrations/2023_12_11_094045_create_contas_imoveis_table.php
@@ -17,14 +17,14 @@ class CreateContasImoveisTable extends Migration
             
             $table->id();
             $table->double('valor');
-            $table->integer('imovelcodigo');
+            $table->unsignedBigInteger('imovelcodigo');
             $table->integer('ano');
             $table->integer('mes');
             $table->string('dataVencimento', 9);
             $table->string('observacoes', 200)->nullable();
             $table->integer('referenciaConta');
             $table->string('nrDocumento', 25)->nullable();
-            $table->integer('salacodigo');
+            $table->unsignedBigInteger('salacodigo');
             $table->timestamps();
 
             $table->foreign('salacodigo')->references('id')->on('salas');
